+++ python/rateslib/curves/curves.py
@@ -25,16 +25,13 @@ from rateslib.calendars.dcfs import _DCF1d
 from rateslib.calendars.rs import CalTypes, get_calendar
 from rateslib.default import NoInput, PlotOutput, _drb, _validate_states, _WithState, plot
 from rateslib.dual import (  # type: ignore[attr-defined]
-    Arr1dF64,
-    Arr1dObj,
     Dual,
     Dual2,
-    DualTypes,
-    Number,
     dual_exp,
     dual_log,
     set_order_convert,
 )
+from rateslib.dual.utils import Arr1dF64, Arr1dObj, DualTypes, Number
 from rateslib.rs import Modifier, index_left_f64
 from rateslib.rs import from_json as from_json_rs
 from rateslib.splines import PPSplineDual, PPSplineDual2, PPSplineF64
